Better error handling(middleware + domain exceptions)

+++ Domain/Services/GraphService.cs
@@ -3,6 +3,7 @@ using GraphCalc.App;
 using GraphCalc.Domain.Entities;
 using GraphCalc.Domain.Interfaces;
 using GraphCalc.Domain.ValueObjects;
+using GraphCalc.Domain.Exceptions;
 
 namespace GraphCalc.Domain.Services;
 
@@ -25,18 +26,44 @@ internal class GraphService : IGraphService
 
     public Graph GetGraphById(Guid id)
     {
-        var graph = graphRepository.GetById(id);
-        return graph;
+        if (id == Guid.Empty)
+            throw new ArgumentException("Graph ID cannot be empty.", nameof(id));
+
+        try
+        {
+            var graph = graphRepository.GetById(id);
+            return graph;
+        }
+        catch (KeyNotFoundException)
+        {
+            throw new EntityNotFoundException(nameof(Graph), id);
+        }
     }
 
     public Graph CreateGraph(Graph graph)
     {
+        ArgumentNullException.ThrowIfNull(graph);
         graphRepository.Add(graph);
         return graph;
     }
 
     public Graph UpdateGraph(Guid id, Graph graph)
     {
+        if (id == Guid.Empty)
+            throw new ArgumentException("Graph ID cannot be empty.", nameof(id));
+
+        ArgumentNullException.ThrowIfNull(graph);
+
+        // Проверяем, что граф существует
+        try
+        {
+            graphRepository.GetById(id);
+        }
+        catch (KeyNotFoundException)
+        {
+            throw new EntityNotFoundException(nameof(Graph), id);
+        }
+
         // ensure the domain model has the correct id
         var updatedGraph = Graph.CreateWithId(id, graph.Range, graph.AuthorId, graph.Items.ToList());
         graphRepository.Update(updatedGraph);
@@ -45,18 +72,39 @@ internal class GraphService : IGraphService
 
     public void DeleteGraph(Guid id)
     {
-        graphRepository.Delete(id);
+        if (id == Guid.Empty)
+            throw new ArgumentException("Graph ID cannot be empty.", nameof(id));
+
+        try
+        {
+            graphRepository.Delete(id);
+        }
+        catch (KeyNotFoundException)
+        {
+            throw new EntityNotFoundException(nameof(Graph), id);
+        }
     }
 
     public List<Series> CalculateGraph(Guid id)
     {
-        var graph = graphRepository.GetById(id);
-        var response = graphService.Calculate(graph);
-        return response;
+        if (id == Guid.Empty)
+            throw new ArgumentException("Graph ID cannot be empty.", nameof(id));
+
+        try
+        {
+            var graph = graphRepository.GetById(id);
+            var response = graphService.Calculate(graph);
+            return response;
+        }
+        catch (KeyNotFoundException)
+        {
+            throw new EntityNotFoundException(nameof(Graph), id);
+        }
     }
 
     public List<Series> CalculateGraph(Graph graph)
     {
+        ArgumentNullException.ThrowIfNull(graph);
         var response = graphService.Calculate(graph);
         return response;
     }
